refactor(skeleton): replacer
